+++ Sodaq_Universal_Tracker/3GNetwork.h
@@ -54,7 +54,7 @@ public:
     */
     bool setActive(bool on);
 
-    uint8_t transmit(uint8_t* buffer, uint8_t size, uint32_t rxTimeout);
+    size_t transmit(uint8_t* buffer, size_t size, uint32_t rxTimeout);
 
     void loopHandler();
